still can't persist local. out of ideas. goodbye cruel world
