+++ notes/src/reducers/index.js
@@ -1,4 +1,5 @@
 import * as act from "../actions";
+import { newSort, local } from "../actions";
 // import Fuse from "fuse.js";
 
 const initialState = {
@@ -19,6 +20,11 @@ const rootReducer = (state = initialState, action) => {
     case act.FETCHING:
       return { ...state, fetchingNotes: true };
     case act.FETCHED:
+      if (local) {
+        let server = [...action.payload];
+        let sorted = newSort(server, local);
+        localStorage.setItem("state", JSON.stringify(sorted));
+      }
       return {
         ...state,
         fetchingNotes: false,
@@ -28,6 +34,8 @@ const rootReducer = (state = initialState, action) => {
     case act.SAVING:
       return { ...state, savingNote: true };
     case act.SAVED:
+      let saved = [...state.notes, action.payload];
+      localStorage.setItem("state", JSON.stringify(saved));
       return {
         ...state,
         savingNote: false,
@@ -38,6 +46,7 @@ const rootReducer = (state = initialState, action) => {
       return { ...state, deletingNote: true };
     case act.DELETED:
       const deleted = state.notes.filter(note => note._id !== action.payload);
+      localStorage.setItem("state", JSON.stringify(deleted));
       return {
         ...state,
         deletingNote: false,
@@ -62,6 +71,7 @@ const rootReducer = (state = initialState, action) => {
         }
         return note;
       });
+      localStorage.setItem("state", JSON.stringify(updated));
       return { ...state, notes: updated };
     // case act.FILTER:
     //   const options = {
